fix recordings having wrong file extension
rec_name was called before url(fmt) so the returned name always had
the default file extension. We fixed it by explicitly setting self.fmt
from both record and play.

This in turn allowed us to turn url into a property as it shouldn't
concern itself with setting the format in the first place.